Fix directory structure so that setuptools works.
