+++ setup.cfg
@@ -1,5 +1,2 @@
 [aliases]
 test=pytest
-
-[tool:pytest]
-norecursedirs=.cache .eggs .idea dist venv *egg*
