--- src/camera/__init__.py
@@ -1,4 +1,0 @@
-from camera.factory import CameraFactory
-from camera.opencv import OpenCVCamera
-
-CameraFactory.register('opencv', OpenCVCamera)
