+++ src/pisurvl/camera/__init__.py
@@ -1,0 +1,5 @@
+from pisurvl.camera.opencv import OpenCVCamera
+
+from pisurvl.camera.factory import CameraFactory
+
+CameraFactory.register('opencv', OpenCVCamera)
