+++ src/pisurvl/motion/inspector.py
